part4 exercise 4.21 - 4.22

+++ part4/controllers/blogs.js
@@ -1,7 +1,6 @@
 const blogsRouter = require('express').Router();
 const Blog = require('../models/blog');
 const User = require('../models/user');
-const jwt = require('jsonwebtoken');
 
 blogsRouter.get('/', async (request, response) => {
   const blogs = await Blog.find({}).populate('user', {
@@ -17,23 +16,20 @@ blogsRouter.post('/', async (request, response) => {
   if (!body.title || !body.url) {
     return response.status(400).end();
   }
-  const token = request.token;
-  const decodedToken = jwt.verify(token, process.env.SECRET);
-  if (!decodedToken.id) {
-    return response.status(401).end();
-  }
-  const user = await User.findById(decodedToken.id);
+  const requestUser = request.user;
+  const userFromDb = await User.findById(requestUser.id);
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes === undefined ? 0 : body.likes,
-    user: user._id,
+    user: requestUser.id,
   });
 
   const savedBlog = await blog.save();
-  user.blogs = user.blogs.concat(savedBlog._id);
-  await user.save();
+  userFromDb.blogs = userFromDb.blogs.concat(savedBlog._id);
+  await userFromDb.save();
   response.status(201).json(savedBlog);
 });
 
@@ -49,7 +45,12 @@ blogsRouter.get('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const user = request.user;
   const blogId = request.params.id;
+  const blogToDeleted = await Blog.findById(blogId);
+  if (blogToDeleted.user.toString() !== user.id.toString()) {
+    return response.status(403).json({ error: 'Not authorized' });
+  }
   await Blog.findByIdAndRemove(blogId);
   response.status(204).end();
 });
